move schemas out of util
